+++ meteoio/ResamplingAlgorithms2D.cc
@@ -30,8 +30,8 @@ namespace mio {
 const Grid2DObject ResamplingAlgorithms2D::BilinearResampling(const Grid2DObject &i_grid, const double &factor)
 {
 	const double cellsize = i_grid.cellsize/factor;
-	const size_t ncols = (size_t)Optim::round( i_grid.ncols*factor );
-	const size_t nrows = (size_t)Optim::round( i_grid.nrows*factor );
+	const size_t ncols = static_cast<size_t>(Optim::round( static_cast<double>(i_grid.ncols)*factor ));
+	const size_t nrows = static_cast<size_t>(Optim::round( static_cast<double>(i_grid.nrows)*factor ));
 	Grid2DObject o_grid(ncols, nrows, cellsize, i_grid.llcorner);
 
 	Bilinear(o_grid, i_grid); //GridObjects always keep nodata
@@ -41,8 +41,8 @@ const Grid2DObject ResamplingAlgorithms2D::BilinearResampling(const Grid2DObject
 const Grid2DObject ResamplingAlgorithms2D::cubicBSplineResampling(const Grid2DObject &i_grid, const double &factor)
 {
 	const double cellsize = i_grid.cellsize/factor;
-	const size_t ncols = (size_t)Optim::round( i_grid.ncols*factor );
-	const size_t nrows = (size_t)Optim::round( i_grid.nrows*factor );
+	const size_t ncols = static_cast<size_t>(Optim::round( static_cast<double>(i_grid.ncols)*factor ));
+	const size_t nrows = static_cast<size_t>(Optim::round( static_cast<double>(i_grid.nrows)*factor ));
 	Grid2DObject o_grid(ncols, nrows, cellsize, i_grid.llcorner);
 
 	cubicBSpline(o_grid, i_grid); //GridObjects always keep nodata
@@ -52,8 +52,8 @@ const Grid2DObject ResamplingAlgorithms2D::cubicBSplineResampling(const Grid2DOb
 const Grid2DObject ResamplingAlgorithms2D::NearestNeighbour(const Grid2DObject &i_grid, const double &factor)
 {
 	const double cellsize = i_grid.cellsize/factor;
-	const size_t ncols = (size_t)Optim::round( i_grid.ncols*factor );
-	const size_t nrows = (size_t)Optim::round( i_grid.nrows*factor );
+	const size_t ncols = static_cast<size_t>(Optim::round( static_cast<double>(i_grid.ncols)*factor ));
+	const size_t nrows = static_cast<size_t>(Optim::round( static_cast<double>(i_grid.nrows)*factor ));
 	Grid2DObject o_grid(ncols, nrows, cellsize, i_grid.llcorner);
 
 	NearestNeighbour(o_grid, i_grid); //GridObjects always keep nodata
